+++ packages/components/legend/doc/make_graph.sh
@@ -1,0 +1,3 @@
+#! /usr/bin/env bash
+
+dot -Tpng ./brushing.dot > ./brushing.png;