+++ packages/components/legend/src/index.js
@@ -1,0 +1,2 @@
+export { default as ColorBinsG } from './ColorBinsG.svelte';
+export { default as ColorBinsDiv } from './ColorBinsDiv.svelte';