+++ packages/docs/site/src/routes/components/_examples/index.js
@@ -1,3 +1,4 @@
 export {default as barchart} from './barchart';
 export {default as choropleth} from './choropleth';
 export {default as histogram} from './histogram';
+export {default as legend} from './legend';