+++ packages/docs/site/src/routes/components/index.json.js
@@ -6,6 +6,7 @@ export function get(req, res) {
 	res.end(JSON.stringify([
 		{name: 'BarchartVDiv', slug: 'BarchartVDiv'},
 		{name: 'ChoroplethG', slug: 'ChoroplethG'},
+		{name: 'ColorScaleG', slug: 'ColorScaleG'},
 		{name: 'HistogramG', slug: 'HistogramG'},
 	]));
 }